requirements.txt
google-generativeai
semanticscholar
